requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask>=1.0.2

# Flask Packages
Flask-Login>=0.4.0
Flask-Migrate>=2.0.2
Flask-Script>=2.0.5
Flask-SQLAlchemy>=2.4.0
Flask-WTF>=0.14.2
Flask-User>=1.0.1.5
Flask-HTTPAuth>=3.2.3
Flask-marshmallow >= 0.12.0
marshmallow-sqlalchemy >= 0.23.1

# JWT
PyJWT>=1.7.1

# Automated tests
pytest>=3.0.5
pytest-cov>=2.4.0